refactor: MovieUiModel 상영일 변수 분리
