+++ app/src/main/java/woowacourse/movie/movies/MovieUiModel.kt
@@ -3,19 +3,31 @@ package woowacourse.movie.movies
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 import woowacourse.movie.domain.Movie
+import java.time.LocalDate
+import java.time.format.DateTimeFormatter
 
 @Parcelize
 data class MovieUiModel(
     val title: String,
-    val periodText: String,
+    val startDate: LocalDate,
+    val endDate: LocalDate,
     val runningTimeText: String,
     val posterResId: Int,
 ) : Parcelable
 
-fun Movie.toMovieUiModel(): MovieUiModel =
-    MovieUiModel(
-        title = this.title,
-        periodText = "$startDate ~ $endDate",
-        runningTimeText = "${runningTime}분",
-        posterResId = this.poster,
-    )
+fun MovieUiModel.periodText(): String {
+    val formatter = DateTimeFormatter.ofPattern("yyyy.MM.dd")
+    return "${startDate.format(formatter)} ~ ${endDate.format(formatter)}"
+}
+
+fun Movie.toMovieUiModel(): MovieUiModel {
+    val movieUiModel =
+        MovieUiModel(
+            title = this.title,
+            startDate = this.startDate,
+            endDate = this.endDate,
+            runningTimeText = "${runningTime}분",
+            posterResId = this.poster,
+        )
+    return movieUiModel
+}
