+++ app/src/main/java/woowacourse/movie/movies/MoviesContract.kt
@@ -1,8 +1,5 @@
 package woowacourse.movie.movies
 
-import android.util.Log
-import woowacourse.movie.domain.movies
-
 interface MoviesContract {
     interface View {
         fun showMovies(uiModels: List<MovieUiModel>) {}
@@ -14,19 +11,5 @@ interface MoviesContract {
         fun loadMovies() {}
 
         fun onMovieSelected(movieUiModel: MovieUiModel) {}
-    }
-}
-
-class MoviesPresenter(
-    private val view: MoviesContract.View,
-) : MoviesContract.Presenter {
-    override fun loadMovies() {
-        val uiModels = movies.map { it.toMovieUiModel() }
-        view.showMovies(uiModels)
-    }
-
-    override fun onMovieSelected(movieUiModel: MovieUiModel) {
-        Log.d("MoviesPresenter", "onMovieSelected for ${movieUiModel.title}")
-        view.navigateToBookingDetail(movieUiModel)
     }
 }
